migrate: graph part 2
