+++ tensorboard/plugins/graph/polymer3/tf_graph/BUILD
@@ -15,18 +15,11 @@ tf_ts_library(
     strict_checks = False,
     tags = ["manual"],
     deps = [
+        "//tensorboard/components_polymer3/polymer:irons_and_papers",
         "//tensorboard/components_polymer3/polymer:legacy_element_mixin",
         "//tensorboard/components_polymer3/tf_dashboard_common",
         "//tensorboard/plugins/graph/polymer3/tf_graph_common",
         "@npm//@polymer/decorators",
-        "@npm//@polymer/iron-flex-layout",
-        "@npm//@polymer/iron-icons",
-        "@npm//@polymer/paper-button",
-        "@npm//@polymer/paper-dropdown-menu",
-        "@npm//@polymer/paper-input",
-        "@npm//@polymer/paper-radio-group",
-        "@npm//@polymer/paper-toggle-button",
-        "@npm//@polymer/paper-tooltip",
         "@npm//@polymer/polymer",
         "@npm//@types/d3",
         "@npm//@types/lodash",
