+++ tensorboard/plugins/graph/polymer3/tf_graph_board/tf-graph-board.ts
@@ -14,8 +14,8 @@ limitations under the License.
 ==============================================================================*/
 import {PolymerElement, html} from '@polymer/polymer';
 import {customElement, observe, property} from '@polymer/decorators';
-import '@polymer/paper-progress';
 
+import '../../../../components_polymer3/polymer/irons_and_papers';
 import '../tf_graph_info/tf-graph-info';
 import '../tf_graph/tf-graph';
 import * as tf_graph from '../tf_graph_common/graph';
