+++ tensorboard/plugins/graph/polymer3/tf_graph_common/scene.ts
@@ -14,8 +14,14 @@ limitations under the License.
 ==============================================================================*/
 import * as d3 from 'd3';
 import * as _ from 'lodash';
-import * as PolymerDom from '@polymer/polymer/lib/legacy/polymer.dom.js';
+import * as PolymerDom from '../../../../components_polymer3/polymer/dom';
 
+import {
+  Class as _Class,
+  selectChild as _selectChild,
+  selectOrCreateChild,
+  SVG_NAMESPACE,
+} from './common';
 import * as edge from './edge';
 import {NodeType, OpNode} from './graph';
 import * as layout from './layout';
@@ -23,64 +29,9 @@ import * as node from './node';
 import * as render from './render';
 import {TfGraphScene} from './tf-graph-scene';
 
-export const SVG_NAMESPACE = 'http://www.w3.org/2000/svg';
-/** Enums element class of objects in the scene */
-export let Class = {
-  Node: {
-    // <g> element that contains nodes.
-    CONTAINER: 'nodes',
-    // <g> element that contains detail about a node.
-    GROUP: 'node',
-    // <g> element that contains visual elements (like rect, ellipse).
-    SHAPE: 'nodeshape',
-    // <*> element(s) under SHAPE that should receive color updates.
-    COLOR_TARGET: 'nodecolortarget',
-    // <text> element showing the node's label.
-    LABEL: 'nodelabel',
-    // <g> element that contains all visuals for the expand/collapse
-    // button for expandable group nodes.
-    BUTTON_CONTAINER: 'buttoncontainer',
-    // <circle> element that surrounds expand/collapse buttons.
-    BUTTON_CIRCLE: 'buttoncircle',
-    // <path> element of the expand button.
-    EXPAND_BUTTON: 'expandbutton',
-    // <path> element of the collapse button.
-    COLLAPSE_BUTTON: 'collapsebutton',
-  },
-  Edge: {
-    CONTAINER: 'edges',
-    GROUP: 'edge',
-    LINE: 'edgeline',
-    REFERENCE_EDGE: 'referenceedge',
-    REF_LINE: 'refline',
-    SELECTABLE: 'selectableedge',
-    SELECTED: 'selectededge',
-    STRUCTURAL: 'structural',
-  },
-  Annotation: {
-    OUTBOX: 'out-annotations',
-    INBOX: 'in-annotations',
-    GROUP: 'annotation',
-    NODE: 'annotation-node',
-    EDGE: 'annotation-edge',
-    CONTROL_EDGE: 'annotation-control-edge',
-    LABEL: 'annotation-label',
-    ELLIPSIS: 'annotation-ellipsis',
-  },
-  Scene: {
-    GROUP: 'scene',
-    CORE: 'core',
-    FUNCTION_LIBRARY: 'function-library',
-    INEXTRACT: 'in-extract',
-    OUTEXTRACT: 'out-extract',
-  },
-  Subscene: {GROUP: 'subscene'},
-  OPNODE: 'op',
-  METANODE: 'meta',
-  SERIESNODE: 'series',
-  BRIDGENODE: 'bridge',
-  ELLIPSISNODE: 'ellipsis',
-};
+export const selectChild = _selectChild;
+export const Class = _Class;
+
 /**
  * The dimensions of the minimap including padding and margin.
  */
@@ -256,208 +207,13 @@ export function panToNode(nodeName: String, svg, zoomG, d3zoom): boolean {
   return false;
 }
 /**
- * Given a container d3 selection, select a child svg element of a given tag
- * and class if exists or append / insert one otherwise.  If multiple children
- * matches the tag and class name, returns only the first one.
- *
- * @param container
- * @param tagName tag name.
- * @param className (optional) Class name or a list of class names.
- * @param before (optional) reference DOM node for insertion.
- * @return selection of the element
- */
-export function selectOrCreateChild(
-  container,
-  tagName: string,
-  className?: string | string[],
-  before?
-): d3.Selection<any, any, any, any> {
-  let child = selectChild(container, tagName, className);
-  if (!child.empty()) {
-    return child;
-  }
-  let newElement = document.createElementNS(
-    'http://www.w3.org/2000/svg',
-    tagName
-  );
-  if (className instanceof Array) {
-    for (let i = 0; i < className.length; i++) {
-      newElement.classList.add(className[i]);
-    }
-  } else {
-    newElement.classList.add(className);
-  }
-  if (before) {
-    // if before exists, insert
-    container.node().insertBefore(newElement, before);
-  } else {
-    // otherwise, append
-    container.node().appendChild(newElement);
-  }
-  return (
-    d3
-      .select(newElement)
-      // need to bind data to emulate d3_selection.append
-      .datum(container.datum())
-  );
-}
-/**
- * Given a container d3 selection, select a child element of a given tag and
- * class. If multiple children matches the tag and class name, returns only
- * the first one.
- *
- * @param container
- * @param tagName tag name.
- * @param className (optional) Class name or list of class names.
- * @return selection of the element, or an empty selection
- */
-export function selectChild(
-  container,
-  tagName: string,
-  className?: string | string[]
-): d3.Selection<any, any, any, any> {
-  let children = container.node().childNodes;
-  for (let i = 0; i < children.length; i++) {
-    let child = children[i];
-    if (child.tagName === tagName) {
-      if (className instanceof Array) {
-        let hasAllClasses = true;
-        for (let j = 0; j < className.length; j++) {
-          hasAllClasses =
-            hasAllClasses && child.classList.contains(className[j]);
-        }
-        if (hasAllClasses) {
-          return d3.select(child);
-        }
-      } else if (!className || child.classList.contains(className)) {
-        return d3.select(child);
-      }
-    }
-  }
-  return d3.select(null);
-}
-/**
- * Select or create a sceneGroup and build/update its nodes and edges.
- *
- * Structure Pattern:
- *
- * <g class='scene'>
- *   <g class='core'>
- *     <g class='edges'>
- *       ... stuff from tf.graph.scene.edges.build ...
- *     </g>
- *     <g class='nodes'>
- *       ... stuff from tf.graph.scene.nodes.build ...
- *     </g>
- *   </g>
- *   <g class='in-extract'>
- *     <g class='nodes'>
- *       ... stuff from tf.graph.scene.nodes.build ...
- *     </g>
- *   </g>
- *   <g class='out-extract'>
- *     <g class='nodes'>
- *       ... stuff from tf.graph.scene.nodes.build ...
- *     </g>
- *   </g>
- * </g>
- *
- * @param container D3 selection of the parent.
- * @param renderNode render node of a metanode or series node.
- * @param sceneElement <tf-graph-scene> polymer element.
- * @param sceneClass class attribute of the scene (default='scene').
- */
-export function buildGroup(
-  container,
-  renderNode: render.RenderGroupNodeInfo,
-  sceneElement: TfGraphScene,
-  sceneClass?: string
-): d3.Selection<any, any, any, any> {
-  sceneClass = sceneClass || Class.Scene.GROUP;
-  let isNewSceneGroup = selectChild(container, 'g', sceneClass).empty();
-  let sceneGroup = selectOrCreateChild(container, 'g', sceneClass);
-  // core
-  let coreGroup = selectOrCreateChild(sceneGroup, 'g', Class.Scene.CORE);
-  let coreNodes = _.reduce(
-    renderNode.coreGraph.nodes(),
-    (nodes, name) => {
-      let node = renderNode.coreGraph.node(name);
-      if (!node.excluded) {
-        nodes.push(node);
-      }
-      return nodes;
-    },
-    []
-  );
-  if (renderNode.node.type === NodeType.SERIES) {
-    // For series, we want the first item on top, so reverse the array so
-    // the first item in the series becomes last item in the top, and thus
-    // is rendered on the top.
-    coreNodes.reverse();
-  }
-  // Create the layer of edges for this scene (paths).
-  edge.buildGroup(coreGroup, renderNode.coreGraph, sceneElement);
-  // Create the layer of nodes for this scene (ellipses, rects etc).
-  node.buildGroup(coreGroup, coreNodes, sceneElement);
-  // In-extract
-  if (renderNode.isolatedInExtract.length > 0) {
-    let inExtractGroup = selectOrCreateChild(
-      sceneGroup,
-      'g',
-      Class.Scene.INEXTRACT
-    );
-    node.buildGroup(inExtractGroup, renderNode.isolatedInExtract, sceneElement);
-  } else {
-    selectChild(sceneGroup, 'g', Class.Scene.INEXTRACT).remove();
-  }
-  // Out-extract
-  if (renderNode.isolatedOutExtract.length > 0) {
-    let outExtractGroup = selectOrCreateChild(
-      sceneGroup,
-      'g',
-      Class.Scene.OUTEXTRACT
-    );
-    node.buildGroup(
-      outExtractGroup,
-      renderNode.isolatedOutExtract,
-      sceneElement
-    );
-  } else {
-    selectChild(sceneGroup, 'g', Class.Scene.OUTEXTRACT).remove();
-  }
-  // Library functions
-  if (renderNode.libraryFunctionsExtract.length > 0) {
-    let outExtractGroup = selectOrCreateChild(
-      sceneGroup,
-      'g',
-      Class.Scene.FUNCTION_LIBRARY
-    );
-    node.buildGroup(
-      outExtractGroup,
-      renderNode.libraryFunctionsExtract,
-      sceneElement
-    );
-  } else {
-    selectChild(sceneGroup, 'g', Class.Scene.FUNCTION_LIBRARY).remove();
-  }
-  position(sceneGroup, renderNode);
-  // Fade in the scene group if it didn't already exist.
-  if (isNewSceneGroup) {
-    sceneGroup
-      .attr('opacity', 0)
-      .transition()
-      .attr('opacity', 1);
-  }
-  return sceneGroup;
-}
-/**
  * Given a scene's svg group, set  g.in-extract, g.coreGraph, g.out-extract svg
  * groups' position relative to the scene.
  *
  * @param sceneGroup
  * @param renderNode render node of a metanode or series node.
  */
-function position(sceneGroup, renderNode: render.RenderGroupNodeInfo) {
+export function position(sceneGroup, renderNode: render.RenderGroupNodeInfo) {
   // Translate scenes down by the label height so that when showing graphs in
   // expanded metanodes, the graphs are below the labels.  Do not shift them
   // down for series nodes as series nodes don't have labels inside of their
@@ -880,7 +636,9 @@ export function addHealthPill(
   statsSvg.setAttribute('x', String(textXOffset));
   statsSvg.setAttribute('y', String(healthPillYOffset - 2));
   healthPillGroup.appendChild(statsSvg);
-  PolymerDom.dom(nodeGroupElement.parentNode).appendChild(healthPillGroup);
+  (PolymerDom.dom(nodeGroupElement.parentNode) as any).appendChild(
+    healthPillGroup
+  );
 }
 /**
  * Adds health pills (which visualize tensor summaries) to a graph group.
