+++ tensorboard/plugins/graph/polymer3/tf_graph_common/tf-graph-scene.ts
@@ -12,6 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
+import * as d3 from 'd3';
 import {RenderNodeInfo} from './render';
 
 type Selection = d3.Selection<any, any, any, any>;
