+++ tensorboard/plugins/graph/polymer3/tf_graph_controls/BUILD
@@ -12,20 +12,12 @@ tf_ts_library(
     strict_checks = False,
     tags = ["manual"],
     deps = [
+        "//tensorboard/components_polymer3/polymer:irons_and_papers",
         "//tensorboard/components_polymer3/polymer:legacy_element_mixin",
         "//tensorboard/components_polymer3/tf_dashboard_common",
         "//tensorboard/plugins/graph/polymer3/tf_graph_common",
         "//tensorboard/plugins/graph/polymer3/tf_graph_node_search",
         "@npm//@polymer/decorators",
-        "@npm//@polymer/iron-icon",
-        "@npm//@polymer/paper-button",
-        "@npm//@polymer/paper-dropdown-menu",
-        "@npm//@polymer/paper-icon-button",
-        "@npm//@polymer/paper-item",
-        "@npm//@polymer/paper-listbox",
-        "@npm//@polymer/paper-radio-group",
-        "@npm//@polymer/paper-toggle-button",
-        "@npm//@polymer/paper-tooltip",
         "@npm//@polymer/polymer",
         "@npm//@types/lodash",
         "@npm//lodash",
