+++ tensorboard/plugins/graph/polymer3/tf_graph_debugger_data_card/BUILD
@@ -12,14 +12,12 @@ tf_ts_library(
     strict_checks = False,
     tags = ["manual"],
     deps = [
+        "//tensorboard/components_polymer3/polymer:dom",
+        "//tensorboard/components_polymer3/polymer:irons_and_papers",
         "//tensorboard/components_polymer3/polymer:legacy_element_mixin",
         "//tensorboard/components_polymer3/tf_dashboard_common",
         "//tensorboard/plugins/graph/polymer3/tf_graph_common",
         "@npm//@polymer/decorators",
-        "@npm//@polymer/paper-material",
-        "@npm//@polymer/paper-slider",
-        "@npm//@polymer/paper-spinner",
-        "@npm//@polymer/paper-toggle-button",
         "@npm//@polymer/polymer",
     ],
 )
