+++ tensorboard/plugins/graph/polymer3/tf_graph_info/BUILD
@@ -14,6 +14,8 @@ tf_ts_library(
     strict_checks = False,
     tags = ["manual"],
     deps = [
+        "//tensorboard/components_polymer3/polymer:dom",
+        "//tensorboard/components_polymer3/polymer:irons_and_papers",
         "//tensorboard/components_polymer3/polymer:legacy_element_mixin",
         "//tensorboard/components_polymer3/tf_dashboard_common",
         "//tensorboard/components_polymer3/tf_wbr_string",
@@ -21,14 +23,6 @@ tf_ts_library(
         "//tensorboard/plugins/graph/polymer3/tf_graph_debugger_data_card",
         "//tensorboard/plugins/graph/polymer3/tf_graph_op_compat_card",
         "@npm//@polymer/decorators",
-        "@npm//@polymer/iron-collapse",
-        "@npm//@polymer/iron-list",
-        "@npm//@polymer/paper-button",
-        "@npm//@polymer/paper-icon-button",
-        "@npm//@polymer/paper-item",
-        "@npm//@polymer/paper-material",
-        "@npm//@polymer/paper-slider",
-        "@npm//@polymer/paper-spinner",
         "@npm//@polymer/polymer",
         "@npm//@types/lodash",
         "@npm//lodash",
