+++ tensorboard/plugins/graph/polymer3/tf_graph_node_search/BUILD
@@ -11,12 +11,11 @@ tf_ts_library(
     tags = ["manual"],
     visibility = ["//visibility:public"],
     deps = [
+        "//tensorboard/components_polymer3/polymer:irons_and_papers",
         "//tensorboard/components_polymer3/polymer:legacy_element_mixin",
         "//tensorboard/components_polymer3/tf_dashboard_common",
         "//tensorboard/plugins/graph/polymer3/tf_graph_common",
         "@npm//@polymer/decorators",
-        "@npm//@polymer/paper-input",
-        "@npm//@polymer/paper-styles",
         "@npm//@polymer/polymer",
         "@npm//@types/lodash",
         "@npm//lodash",
