+++ tensorboard/plugins/graph/polymer3/tf_graph_op_compat_card/BUILD
@@ -13,14 +13,11 @@ tf_ts_library(
     strict_checks = False,
     tags = ["manual"],
     deps = [
+        "//tensorboard/components_polymer3/polymer:irons_and_papers",
         "//tensorboard/components_polymer3/polymer:legacy_element_mixin",
         "//tensorboard/components_polymer3/tf_dashboard_common",
         "//tensorboard/plugins/graph/polymer3/tf_graph_common",
         "@npm//@polymer/decorators",
-        "@npm//@polymer/iron-collapse",
-        "@npm//@polymer/iron-list",
-        "@npm//@polymer/paper-icon-button",
-        "@npm//@polymer/paper-item",
         "@npm//@polymer/polymer",
         "@npm//@types/d3",
         "@npm//d3",
